Save item in slot with Middle click, clear with Ctrl+Middle
In order to use this functionality, put desired item in selected guild
vault slot and press it with middle mouse button. This is equivalent
to calling `/gvs set <slot>:<item>` chat command on the clicked slot.

It is also possible to clear slot assignment; press the desired slot
with Ctrl+Middle click to remove item assignment. This is equivalent
to calling `/gvs clear <slot>`.